feat(theme): introduce isolated paletteNext and foundationNext
